Bug 1456394 - Move KeyframeEffect constructors to KeyframeEffectReadOnly.{h,cpp}; r=bz,hiro
By doing this we will have all the KeyframeEffect* related code in KeyframeEffectReadOnly.{h,cpp} so we can rename them to KeyframeEffect.{h,cpp} in the next patch and make it easier to examine the history for the bulk of this code.

The added [HeaderFile] annotation will be removed in a subsequent patch in this series.
