+++ dom/webidl/KeyframeEffect.webidl
@@ -62,6 +62,7 @@ partial interface KeyframeEffectReadOnly {
 // processing on the `keyframes` object.
 [Func="nsDocument::IsWebAnimationsEnabled",
  RunConstructorInCallerCompartment,
+ HeaderFile="mozilla/dom/KeyframeEffectReadOnly.h",
  Constructor ((Element or CSSPseudoElement)? target,
               object? keyframes,
               optional (unrestricted double or KeyframeEffectOptions) options),
